refactor(Naming): 통일된 구조를 위한 이름 변경
- Rank라는 단어가 기존 구조와 통일적이지 못하다고 판단
- 따라서 Rank -> TotalResult로 변경
